+++ src/theory/arith/nl/nonlinear_extension.h
@@ -15,8 +15,8 @@
  ** multiplication via axiom instantiations.
  **/
 
-#ifndef CVC4__THEORY__ARITH__NONLINEAR_EXTENSION_H
-#define CVC4__THEORY__ARITH__NONLINEAR_EXTENSION_H
+#ifndef CVC4__THEORY__ARITH__NL__NONLINEAR_EXTENSION_H
+#define CVC4__THEORY__ARITH__NL__NONLINEAR_EXTENSION_H
 
 #include <stdint.h>
 #include <map>
@@ -25,11 +25,11 @@
 #include "context/cdlist.h"
 #include "expr/kind.h"
 #include "expr/node.h"
-#include "theory/arith/nl_lemma_utils.h"
-#include "theory/arith/nl_model.h"
-#include "theory/arith/nl_solver.h"
+#include "theory/arith/nl/nl_lemma_utils.h"
+#include "theory/arith/nl/nl_model.h"
+#include "theory/arith/nl/nl_solver.h"
 #include "theory/arith/theory_arith.h"
-#include "theory/arith/transcendental_solver.h"
+#include "theory/arith/nl/transcendental_solver.h"
 #include "theory/uf/equality_engine.h"
 
 namespace CVC4 {
